+++ packages/o-spreadsheet-engine/src/constants.ts
@@ -1,0 +1,1 @@
+export const NEWLINE = "\n";
