+++ packages/o-spreadsheet-engine/src/formulas/tokenizer.ts
@@ -1,0 +1,223 @@
+import { NEWLINE } from "../constants";
+import { TokenizingChars, replaceNewLines, specialWhiteSpaceRegexp } from "../helpers/misc";
+import { getFormulaNumberRegex } from "../helpers/numbers";
+import { rangeReference } from "../helpers/references";
+import { CellErrorType } from "../types/errors";
+import { DEFAULT_LOCALE, type Locale } from "../types/locale";
+
+export const POSTFIX_UNARY_OPERATORS = ["%"];
+const OPERATORS = "+,-,*,/,:,=,<>,>=,>,<=,<,^,&".split(",").concat(POSTFIX_UNARY_OPERATORS);
+
+type TokenType =
+  | "OPERATOR"
+  | "NUMBER"
+  | "STRING"
+  | "SYMBOL"
+  | "SPACE"
+  | "DEBUGGER"
+  | "ARG_SEPARATOR"
+  | "LEFT_PAREN"
+  | "RIGHT_PAREN"
+  | "REFERENCE"
+  | "INVALID_REFERENCE"
+  | "UNKNOWN";
+
+export interface Token {
+  readonly type: TokenType;
+  readonly value: string;
+}
+
+export function tokenize(str: string, locale: Locale = DEFAULT_LOCALE): Token[] {
+  str = replaceNewLines(str);
+  const chars = new TokenizingChars(str);
+  const result: Token[] = [];
+  const tokenizeSpace = specialWhiteSpaceRegexp.test(str)
+    ? tokenizeSpecialCharacterSpace
+    : tokenizeSimpleSpace;
+
+  while (!chars.isOver()) {
+    let token =
+      tokenizeNewLine(chars) ||
+      tokenizeSpace(chars) ||
+      tokenizeArgsSeparator(chars, locale) ||
+      tokenizeParenthesis(chars) ||
+      tokenizeOperator(chars) ||
+      tokenizeString(chars) ||
+      tokenizeDebugger(chars) ||
+      tokenizeInvalidRange(chars) ||
+      tokenizeNumber(chars, locale) ||
+      tokenizeSymbol(chars);
+
+    if (!token) {
+      token = { type: "UNKNOWN", value: chars.shift() };
+    }
+
+    result.push(token);
+  }
+  return result;
+}
+
+function tokenizeDebugger(chars: TokenizingChars): Token | null {
+  if (chars.current === "?") {
+    chars.shift();
+    return { type: "DEBUGGER", value: "?" };
+  }
+  return null;
+}
+
+const parenthesis = {
+  "(": { type: "LEFT_PAREN", value: "(" },
+  ")": { type: "RIGHT_PAREN", value: ")" },
+} as const;
+
+function tokenizeParenthesis(chars: TokenizingChars): Token | null {
+  if (chars.current === "(" || chars.current === ")") {
+    const value = chars.shift();
+    return parenthesis[value];
+  }
+  return null;
+}
+
+function tokenizeArgsSeparator(chars: TokenizingChars, locale: Locale): Token | null {
+  if (chars.current === locale.formulaArgSeparator) {
+    const value = chars.shift();
+    const type = "ARG_SEPARATOR";
+    return { type, value };
+  }
+
+  return null;
+}
+
+function tokenizeOperator(chars: TokenizingChars): Token | null {
+  for (const op of OPERATORS) {
+    if (chars.currentStartsWith(op)) {
+      chars.advanceBy(op.length);
+      return { type: "OPERATOR", value: op };
+    }
+  }
+  return null;
+}
+
+const FIRST_POSSIBLE_NUMBER_CHARS = new Set("0123456789");
+
+function tokenizeNumber(chars: TokenizingChars, locale: Locale): Token | null {
+  if (
+    !FIRST_POSSIBLE_NUMBER_CHARS.has(chars.current) &&
+    chars.current !== locale.decimalSeparator
+  ) {
+    return null;
+  }
+  const match = chars.remaining().match(getFormulaNumberRegex(locale.decimalSeparator));
+  if (match) {
+    chars.advanceBy(match[0].length);
+    return { type: "NUMBER", value: match[0] };
+  }
+  return null;
+}
+
+function tokenizeString(chars: TokenizingChars): Token | null {
+  if (chars.current === '"') {
+    const startChar = chars.shift();
+    let letters: string = startChar;
+    while (chars.current && (chars.current !== startChar || letters[letters.length - 1] === "\\")) {
+      letters += chars.shift();
+    }
+    if (chars.current === '"') {
+      letters += chars.shift();
+    }
+    return {
+      type: "STRING",
+      value: letters,
+    };
+  }
+  return null;
+}
+
+const unicodeSymbolCharRegexp = /\p{L}|\p{N}|_|\.|!|\$/u;
+const SYMBOL_CHARS = new Set("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_.!$");
+
+function tokenizeSymbol(chars: TokenizingChars): Token | null {
+  let result: string = "";
+  if (chars.current === "'") {
+    let lastChar = chars.shift();
+    result += lastChar;
+    while (chars.current) {
+      lastChar = chars.shift();
+      result += lastChar;
+      if (lastChar === "'") {
+        if (chars.current && chars.current === "'") {
+          lastChar = chars.shift();
+          result += lastChar;
+        } else {
+          break;
+        }
+      }
+    }
+
+    if (lastChar !== "'") {
+      return {
+        type: "UNKNOWN",
+        value: result,
+      };
+    }
+  }
+  while (
+    chars.current &&
+    (SYMBOL_CHARS.has(chars.current) || chars.current.match(unicodeSymbolCharRegexp))
+  ) {
+    result += chars.shift();
+  }
+  if (result.length) {
+    const value = result;
+    const isReference = rangeReference.test(value);
+    if (isReference) {
+      return { type: "REFERENCE", value };
+    }
+    return { type: "SYMBOL", value };
+  }
+  return null;
+}
+
+function tokenizeSpecialCharacterSpace(chars: TokenizingChars): Token | null {
+  let spaces = "";
+  while (chars.current === " " || (chars.current && chars.current.match(specialWhiteSpaceRegexp))) {
+    spaces += chars.shift();
+  }
+
+  if (spaces) {
+    return { type: "SPACE", value: spaces };
+  }
+  return null;
+}
+
+function tokenizeSimpleSpace(chars: TokenizingChars): Token | null {
+  let spaces = "";
+  while (chars.current === " ") {
+    spaces += chars.shift();
+  }
+
+  if (spaces) {
+    return { type: "SPACE", value: spaces };
+  }
+  return null;
+}
+
+function tokenizeNewLine(chars: TokenizingChars): Token | null {
+  let length = 0;
+  while (chars.current === NEWLINE) {
+    length++;
+    chars.shift();
+  }
+  if (length) {
+    return { type: "SPACE", value: NEWLINE.repeat(length) };
+  }
+  return null;
+}
+
+function tokenizeInvalidRange(chars: TokenizingChars): Token | null {
+  if (chars.currentStartsWith(CellErrorType.InvalidReference)) {
+    chars.advanceBy(CellErrorType.InvalidReference.length);
+    return { type: "INVALID_REFERENCE", value: CellErrorType.InvalidReference };
+  }
+  return null;
+}
