+++ src/index.ts
@@ -502,5 +502,3 @@ export type { FunctionRegistry };
 
 export { AbstractCellClipboardHandler } from "./clipboard_handlers/abstract_cell_clipboard_handler";
 export { AbstractFigureClipboardHandler } from "./clipboard_handlers/abstract_figure_clipboard_handler";
-
-export * from "@odoo/o-spreadsheet-engine";
